Removed useless classes and methods, code is optimized

+++ ServerForLab/ServerForLab/CipherUtils/PrivateKey.cs
@@ -4,10 +4,10 @@ namespace ServerForLab.RSAcipher
 {
     public class PrivateKey
     {
-        public long D { get; }
-        public long N { get; }
+        public BigInteger D { get; }
+        public BigInteger N { get; }
 
-        public PrivateKey(long d, long n)
+        public PrivateKey(BigInteger d, BigInteger n)
         {
             this.D = d;
             this.N = n;
